requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
langchain==0.2.3
langchain-community==0.2.3
chromadb==0.5.0
sentence-transformers==2.7.0
pypdf==4.2.0
docx2txt==0.8
python-multipart==0.0.9
python-dotenv==1.0.1
tiktoken==0.6.0
pydantic==2.7.1
requests==2.31.0
pytest==8.2.2
httpx==0.27.0
